# docs/swagger.yaml
basePath: /
definitions:
  models.Device:
    properties:
      deviceModel:
        example: /devicemodels/id1
        type: string
      id:
        example: /devices/id1
        type: string
      name:
        example: Sensor
        type: string
      note:
        example: Testing a sensor
        type: string
      serial:
        example: A020000102
        type: string
    required:
    - deviceModel
    - id
    - name
    - note
    - serial
    type: object
  models.Error:
    properties:
      message:
        type: string
    type: object
host: tvwvnaqfy9.execute-api.us-west-2.amazonaws.com/api
info:
  contact:
    email: m.dehghanpour10@gmail.com
    name: Mohammad Dehghanpour
  description: Implement a simple Restful API on AWS
  title: GO challenge
  version: "1.0"
paths:
  /devices:
    post:
      consumes:
      - application/json
      description: this endpoint create a new device info
      parameters:
      - description: device id
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/models.Device'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Device'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: create a new device info
      tags:
      - device
  /devices/{id}:
    get:
      consumes:
      - application/json
      description: this endpoint receives id of device and back dive info
      parameters:
      - description: device id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Device'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: get a device info by id
      tags:
      - device
swagger: "2.0"